+++ LocomotionWebApp/Models/ViewModels/PatientViewModel.cs
@@ -29,9 +29,9 @@ namespace LocomotionWebApp.Models.ViewModels
 
 		public string Deformity { get; set; }
 
-		public int ShankLength { get; set; }
+		public double ShankLength { get; set; }
 
-		public int ThighLength { get; set; }
+		public double ThighLength { get; set; }
 
 		public int Age { get; set; }
 
@@ -39,9 +39,9 @@ namespace LocomotionWebApp.Models.ViewModels
 
 		public string Gender { get; set; }
 
-		public int Height { get; set; }
+		public double Height { get; set; }
 
-		public int Weight { get; set; }
+		public double Weight { get; set; }
 
 		public DateTime Start { get; set; } //When Therapists adds the patient
